+++ src/core.rs
@@ -10,7 +10,8 @@ use crate::{
         update_last_verified,
     },
     db_kind::DbKind,
-    db_value::{ColumnMap, DbParams, DbRow, DbRows, IntoDbParams, IntoDbRows},
+    db_row::{ColumnMap, DbRow, DbRows, IntoDbRows},
+    db_value::{DbParams, IntoDbParams},
     parse::get_accessed_tables,
 };
