+++ src/lib.rs
@@ -4,6 +4,7 @@ pub mod any;
 pub mod cache;
 pub mod core;
 pub mod db_kind;
+pub mod db_row;
 pub mod db_type;
 pub mod db_value;
 pub mod parse;
