+++ src/serde.rs
@@ -2,7 +2,8 @@
 
 use crate::{
     core::DbError,
-    db_value::{DbRow, DbValue, JsonRow, JsonValue},
+    db_row::{DbRow, JsonRow},
+    db_value::{DbValue, JsonValue},
 };
 use serde::{
     Deserialize, Serialize,
