+++ src/shared.rs
@@ -1,7 +1,8 @@
 use crate::{
     cache::clear_cache_for_edited_tables,
     core::{DbError, DbQuery},
-    db_value::{DbRows, DbValue, IntoDbRows},
+    db_row::{DbRows, IntoDbRows},
+    db_value::DbValue,
     parse::validate_table_name,
 };
 use std::fmt::Display;
